requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==3.2.2
cffi==1.17.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
mako==1.3.10
markupsafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pyjwt==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
